requirements.txt
dipy
imageio
matplotlib
nibabel
numpy
pydicom
scikit-image
scipy
SimpleITK
medpy
pillow
plotly